moved ConfigFile class to util package
